Refactor legacy grids - tool shed webapp shouldn't depend on reports webapp.
