+++ lib/galaxy/webapps/reports/controllers/workflows.py
@@ -17,13 +17,13 @@ from markupsafe import escape
 from sqlalchemy import and_
 
 from galaxy import model, util
+from galaxy.web.legacy_framework import grids
 from galaxy.webapps.base.controller import BaseUIController, web
 from galaxy.webapps.reports.controllers.jobs import (
     get_spark_time,
     sorter
 )
 from galaxy.webapps.reports.controllers.query import ReportQueryBuilder
-from galaxy.webapps.reports.framework import grids
 
 log = logging.getLogger(__name__)
