+++ lib/tool_shed/grids/admin_grids.py
@@ -4,7 +4,7 @@ from markupsafe import escape
 from sqlalchemy import and_
 
 from galaxy.web.framework.helpers import time_ago
-from galaxy.webapps.reports.framework import grids
+from galaxy.web.legacy_framework import grids
 from tool_shed.grids.repository_grids import CategoryGrid, RepositoryGrid
 from tool_shed.util import hg_util
 from tool_shed.webapp import model
