+++ lib/tool_shed/grids/repository_review_grids.py
@@ -3,7 +3,7 @@ import logging
 from markupsafe import escape
 from sqlalchemy import and_, false, null, or_, true
 
-from galaxy.webapps.reports.framework import grids
+from galaxy.web.legacy_framework import grids
 from tool_shed.grids.repository_grids import RepositoryGrid
 from tool_shed.util import hg_util, metadata_util
 from tool_shed.webapp import model
