+++ lib/tool_shed/webapp/controllers/repository.py
@@ -22,8 +22,8 @@ from galaxy import (
 )
 from galaxy.tools.repositories import ValidationContext
 from galaxy.web.form_builder import CheckboxField, SelectField
+from galaxy.web.legacy_framework import grids
 from galaxy.webapps.base.controller import BaseUIController
-from galaxy.webapps.reports.framework import grids
 from tool_shed.dependencies.repository import relation_builder
 from tool_shed.galaxy_install import dependency_display
 from tool_shed.metadata import repository_metadata_manager
